+++ inbox/api/wsgi.py
@@ -5,8 +5,8 @@ import gunicorn.glogging
 from gevent.pywsgi import WSGIHandler, WSGIServer
 from gunicorn.workers.ggevent import GeventWorker
 
+from inbox.instrumentation import Tracer
 from inbox.logging import configure_logging, get_logger
-from nylas.util.debug import Tracer
 
 log = get_logger()
