--- nylas/util/__init__.py
@@ -1,4 +1,0 @@
-from pkgutil import extend_path
-
-# Allow out-of-tree submodules.
-__path__ = extend_path(__path__, __name__)
